clientv3/ordering: Split mocked part of the test from integration-level test.
